Maven2Gradle. Added initial coverage. Started removing stuff that calls out to maven on path.
Removed the call to get the effective settings. It was only used to get the uri to the local maven repository. We don't need that because we already have working implementation that locates the local maven repo. Hence, we can simply add 'mavenLocal()' to the repositories section in the generated build.gradle files.